# swagger_schemas/swagger/json/crestdb.json
{
	"swagger": "2.0",
	"info": {
		"version": "2.0",
		"title": "CrestDB REST API",
		"termsOfService": "http://svom.swagger.fr/terms/",
		"description": "Crest Rest Api to manage data for calibration files.",
		"contact": {
			"email": "andrea.formica@cern.ch"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "localhost:8080",
	"basePath": "/crest",
...
	"schemes": [
		"http"
	],
	"paths": {
		"/admin/globaltags/{name}": {
			"put": {
				"tags": [
					"admin"
				],
				"summary": "Update a GlobalTag in the database.",
				"description": "This method allows to update a GlobalTag.Arguments: the name has to uniquely identify a global tag.",
				"operationId": "updateGlobalTag",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "A json string that is used to construct a GlobalTagDto object: { name: xxx, ... }",
						"required": true,
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagDto"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"admin"
				],
				"summary": "Remove a GlobalTag from the database.",
				"description": "This method allows to remove a GlobalTag.Arguments: the name has to uniquely identify a global tag.",
				"operationId": "removeGlobalTag",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"default": {
						"description": "successful operation"
					}
				}
			}
		},
...
		"/folders": {
			"post": {
				"tags": [
					"folders"
				],
				"summary": "Create an entry for folder information.",
				"description": "Folder informations go into a dedicated table.",
				"operationId": "createFolder",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json",
					"text/plain"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "A json string that is used to construct a folderdto object: { node: xxx, ... }",
						"required": true,
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/FolderDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "string"
						}
					}
				}
			},
			"get": {
				"tags": [
					"folders"
				],
				"summary": "Finds a FolderDto list.",
				"description": "This method allows to perform search and sorting.Arguments: by=<pattern>, sort=<sortpattern>. The pattern <pattern> is in the form <param-name><operation><param-value>       <param-name> is the name of one of the fields in the dto       <operation> can be [< : >] ; for string use only [:]        <param-value> depends on the chosen parameter. A list of this criteria can be provided       using comma separated strings for <pattern>.      The pattern <sortpattern> is <field>:[DESC|ASC]",
				"operationId": "listFolders",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "by",
						"in": "query",
						"description": "by: the search pattern {none}",
						"required": false,
						"type": "string",
						"default": "none"
					},
					{
						"name": "sort",
						"in": "query",
						"description": "sort: the sort pattern {nodeFullpath:ASC}",
						"required": false,
						"type": "string",
						"default": "nodeFullpath:ASC"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/FolderSetDto"
						}
					}
				}
			}
		},
...
		"/globaltags": {
			"get": {
				"tags": [
					"globaltags"
				],
				"summary": "Finds a GlobalTagDtos lists.",
				"description": "This method allows to perform search and sorting.Arguments: by=<pattern>, page={ipage}, size={isize}, sort=<sortpattern>. The pattern <pattern> is in the form <param-name><operation><param-value>       <param-name> is the name of one of the fields in the dto       <operation> can be [< : >] ; for string use only [:]        <param-value> depends on the chosen parameter. A list of this criteria can be provided       using comma separated strings for <pattern>.      The pattern <sortpattern> is <field>:[DESC|ASC]",
				"operationId": "listGlobalTags",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "by",
						"in": "query",
						"description": "by: the search pattern {none}",
						"required": false,
						"type": "string",
						"default": "none"
					},
					{
						"name": "page",
						"in": "query",
						"description": "page: the page number {0}",
						"required": false,
						"type": "integer",
						"default": 0,
						"format": "int32"
					},
					{
						"name": "size",
						"in": "query",
						"description": "size: the page size {1000}",
						"required": false,
						"type": "integer",
						"default": 1000,
						"format": "int32"
					},
					{
						"name": "sort",
						"in": "query",
						"description": "sort: the sort pattern {name:ASC}",
						"required": false,
						"type": "string",
						"default": "name:ASC"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagSetDto"
						}
					}
				}
			},
			"post": {
				"tags": [
					"globaltags"
				],
				"summary": "Create a GlobalTag in the database.",
				"description": "This method allows to insert a GlobalTag.Arguments: GlobalTagDto should be provided in the body as a JSON file.",
				"operationId": "createGlobalTag",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "A json string that is used to construct a globaltagdto object: { name: xxx, ... }",
						"required": true,
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagDto"
						}
					},
					{
						"name": "force",
						"in": "query",
						"description": "force: tell the server if it should use or not the insertion time provided {default: false}",
						"required": false,
						"type": "string",
						"default": "false"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagDto"
						}
					}
				}
			}
		},
		"/globaltags/{name}": {
			"get": {
				"tags": [
					"globaltags"
				],
				"summary": "Finds a GlobalTagDto by name",
				"description": "This method will search for a global tag with the given name. Only one global tag should be returned.",
				"operationId": "findGlobalTag",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GlobalTagDto"
						}
					}
				}
			}
		},
		"/globaltags/{name}/tags": {
			"get": {
				"tags": [
					"globaltags"
				],
				"summary": "Finds a TagDtos lists associated to the global tag name in input.",
				"description": "This method allows to trace a global tag.Arguments: record=<record> filter output by record, label=<label> filter output by label",
				"operationId": "findGlobalTagFetchTags",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "record",
						"in": "query",
						"description": "record:  the record string {}",
						"required": false,
						"type": "string",
					    "default": "none"
					},
					{
						"name": "label",
						"in": "query",
						"description": "label:  the label string {}",
						"required": false,
						"type": "string",
					    "default": "none"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/TagSetDto"
						}
					}
				}
			}
		},
...
		"/tags/{name}": {
			"get": {
				"tags": [
					"tags"
				],
				"summary": "Finds a TagDto by name",
				"description": "This method will search for a tag with the given name. Only one tag should be returned.",
				"operationId": "findTag",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"description": "name: the tag name",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/TagSetDto"
						}
					}
				}
			},
			"put": {
				"tags": [
					"tags"
				],
				"summary": "Update a TagDto by name",
				"description": "This method will search for a tag with the given name, and update its content for the provided body fields. Only the following fields can be updated: description, timeType, objectTime, endOfValidity, lastValidatedTime.",
				"operationId": "updateTag",
				"produces": [
					"application/json",
					"application/xml"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"description": "name: the tag name",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "A json string that is used to construct a map of updatable fields: { description: xxx, ... }",
						"required": true,
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/GenericMap"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "../../components/json/crestdata.json#/definitions/TagDto"
						}
					}
				}
			}
		},
...
	}
}